rename step file to annotations
