+++ mavis/test/annotations.py
@@ -1,12 +1,13 @@
 from functools import wraps
 
+import os
 import allure
 from io import BytesIO
 from PIL import Image
 from PIL.Image import Palette
 
 
-def reduce_colors(image_bytes: bytes) -> bytes:
+def _reduce_colors(image_bytes: bytes) -> bytes:
     with BytesIO(image_bytes) as input_io:
         img = Image.open(input_io)
         img = img.convert("P", palette=Palette.ADAPTIVE, colors=32)
@@ -25,7 +26,7 @@ def step(title: str, attach_screenshot: bool = True):
                 except Exception:
                     if attach_screenshot:
                         screenshot_bytes = self.page.screenshot(full_page=True)
-                        reduced_bytes = reduce_colors(screenshot_bytes)
+                        reduced_bytes = _reduce_colors(screenshot_bytes)
                         allure.attach(
                             reduced_bytes,
                             name="Screenshot on failure",
@@ -43,7 +44,7 @@ def step(title: str, attach_screenshot: bool = True):
 
                 if attach_screenshot:
                     screenshot_bytes = self.page.screenshot(full_page=True)
-                    reduced_bytes = reduce_colors(screenshot_bytes)
+                    reduced_bytes = _reduce_colors(screenshot_bytes)
                     allure.attach(
                         reduced_bytes,
                         name="Screenshot",
