+++ mavis/test/pages/log_in.py
@@ -1,7 +1,7 @@
 from playwright.sync_api import Page
 
 from mavis.test.models import Organisation, User
-from mavis.test.step import step
+from mavis.test.annotations import step
 
 
 class LogInPage:
