+++ mavis/test/pages/vaccines.py
@@ -3,7 +3,7 @@ from datetime import date
 from playwright.sync_api import Page, expect
 
 from mavis.test.models import Vaccine
-from mavis.test.step import step
+from mavis.test.annotations import step
 
 
 class BatchExpiryDateMixin:
